CORE-5301 Added "New Window" button to Data window.
Added requestNewView and newViewRequested methods to OpenFolderEvent.
Added a "New Window" button to DiskResourceViewToolbar and a DiskResourcePresenter#onNewWindow method that will fire the OpenFolderEvent.
Updated the ConfigFactory#diskResourceWindowConfig method to accept a newWindowRequested flag, which causes the config to be created with a unique tag, which causes a new window to be opened with that config.
Updated ShowWindowEventHandler#onRequestOpenFolder to pass the OpenFolderEvent#newViewRequested flag to the ConfigFactory#diskResourceWindowConfig method.
Updated ToolbarButtonVisibilityNavigationHandler to allow the "New Window" button to be clicked when any folder (or no folder) is selected.