CORE-5732 Added a clear method to LiveGridCheckBoxSelectionModel.
Updated DiskResourceViewImpl to call this clear method, which will clear the model's selections, every time a new folder is loaded in the center grid.
Set the grid's cache size smaller, to cut down on load times, now that listings take a little longer with favorites icons and the like.
Minor refactor in LiveGridCheckBoxSelectionModel#isSelected (itemKey only needs to be calculated once).