CORE-6309 First app section now selectable in editor
This was caused by a bug in the AppEditor presenter editor visitor event
wiring logic. The argument group selected handler was being added, but
prematurely removed on editor initialization. To be specific, the
handlers were being removed at each execution of
RegisterEventHandlers.visit.

This was corrected by moving the Editor-specific handler un-registering
logic out of the RegisterEventHandlers editor visitor and into the
AppsEditor presenter itself. This is more appropriate since it fits into
the intent of the presenter in MVP, and also keeps the intent of the
editor visitor more concise.