Removed the additivity flag on the metrics logger.
This is necessary because the way volumes are mounted into the UI
container. The permissions are not correct, and the UI was trying to log
all metrics to a separate file. This breaks convention with the way our
logging is handled elsewhere.

Removing this flag will ensure that the metrics logging will go to
STDOUT.