Upgrade to GWT 2.7.0 and GXT 3.1.4
Developers will need to set their Sencha support passwords in their
~/.gradle/gradle.properties file. See the [build.gradle] file for
details in the comments.

Many of the *.gwt.xml files had to have some deprecated declarations
removed.

A dependency to 'org.ow2.asm:asm' had to be added to correct some build
errors for the GWT upgrade.

The `TextButton` in the AppEditorPresenterImpl.java
`PromptForSaveDialog` had to be changed to a GXT TextButton, since the
package that the GWT `TextButton` belonged to has been
[deprecated[(http://www.gwtproject.org/release-notes.html#Release_Notes_2_7_0).

CORE-6972