CORE-6952 Refactored SelectFolderByPathLoadHandler.
Refactored SelectFolderByPathLoadHandler into a LoadHanlder once again, but its constructor is now protected and is instantiated by a static SelectFolderByPathLoadHandler#registerFolderLoader method that will also add it to the given tree loader.
Fixed tests in SelectFolderByPathLoadHandlerTest and updated its comments and docs to once again make sense with the latest changes.