Change dewey to use a queue named indexing.ENV
ENV refers to a descriptive name for the environment it's running in. This should make it possible to run dewey off the same AMQP instance in two environments (i.e., in parasitic environments) as well as making it more plausible to horizontally scale by spinning up extra instances of dewey.