CORE-6052 data search trash filtering
The generated elasticsearch queries contain path filtering logic.
Depending on the value of the search query parameter includeTrash,
queries will include or exclude items from the DE trash for the
authenicated user.

As a side-effect, everything that is not inside a home folder or
possible the DE trash folder will be excluded from search results.

To simplify the logic the zone information is no longer added to the
authenticated username in the terrain.routes.search namespace.