CORE-7496, CORE-7497: create a service for indexing template metadata into elasticsearch
Much of the basic structure of this is copied from jobservices (in
jex-go branch) or from jex-events:

 * configurate, logcabin packages
 * flag parsing, version handling
 * mode-setting logic
 * basic docker interaction

At present, this service only operates in one of the three documented
modes, --mode full. This mode indexes the metadata database's 'avus'
table fully once and does not listen on AMQP.

Vendored dependencies are in a separate commit for readability.