CORE-7511 Add more user info to PermID Request details.
Updated the responses of Permanent ID Request endpoints that return a
request's details to include more user info in the 'requested_by' field,
such as email and first/last name.
Since the requesting user's email is now available when notifications
are sent to that user, notification emails can now be enabled.
Added the new ID to the "Completion" notification update in the
"comments" field, which is now included as the body of the "Completion"
notification email.