Began work on handling stop requests in condor-launcher
This work contains code that extracts a condor ID from a []byte
that is presumed to be the output of the 'condor_q -long` command.
A unit test is provided, more is forthcoming.

This also contains the start of the code to execute the
'condor_q -long` command and grab its output. This is untested,
but that will change in the near future.