Added shell.sh to apps.
Starts up a bash shell inside a discoenv/buildenv container that
has the local apps checkout mounted into it. The container is
linked to the apps-de-db container, which is started if it is not
already running. This allows you to run leiningen commands
manually in a clean environment that can connect to the de db
running in the apps-de-db container.