CORE-7544 Fix reference genome path/name validation.
The `not=` symbol needs to be quoted when used outside of the korma
`where` macro. Otherwise these conditionals were parsed as queries like
`WHERE (? AND "genome_reference"."path" = ?)` which always found a match
when attempting to update just the path or name of an existing reference
genome.
Removed unused apps.metadata.reference-genomes/get-all-reference-genomes